Fix tests that failed because they specifyed a custom page swapper that didn't exist
