+++ community/kernel/src/test/resources/META-INF/services/org.neo4j.io.pagecache.PageSwapperFactory
@@ -1,2 +1,2 @@
 org.neo4j.io.pagecache.impl.SingleFilePageSwapperFactory
-org.neo4j.kernel.impl.pagecache.ConfiguringPageCacheFactoryTest$PageSwapperFactoryForTesting
+org.neo4j.kernel.impl.pagecache.PageSwapperFactoryForTesting
