+++ community/neo4j/src/test/java/upgrade/PlatformConstraintStoreUpgradeTest.java
@@ -37,6 +37,7 @@ import org.neo4j.test.TestGraphDatabaseFactory;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
+import static org.neo4j.kernel.impl.pagecache.PageSwapperFactoryForTesting.TEST_PAGESWAPPER_NAME;
 import static org.neo4j.kernel.impl.storemigration.MigrationTestUtils.prepareSampleLegacyDatabase;
 
 public class PlatformConstraintStoreUpgradeTest
@@ -75,6 +76,6 @@ public class PlatformConstraintStoreUpgradeTest
     {
         return new TestGraphDatabaseFactory().newEmbeddedDatabaseBuilder( workingDir )
                 .setConfig( GraphDatabaseSettings.allow_store_upgrade, "true" )
-                .setConfig( GraphDatabaseSettings.pagecache_swapper, "custom" ).newGraphDatabase();
+                .setConfig( GraphDatabaseSettings.pagecache_swapper, TEST_PAGESWAPPER_NAME ).newGraphDatabase();
     }
 }
